requirements.txt
# Core AI Agent Framework
smolagents>=0.4.0
transformers>=4.35.0
huggingface-hub>=0.19.0
torch>=2.0.0

# LangChain Fallback
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Local LLM Integration
ollama>=0.1.0
requests>=2.31.0

# Web Scraping
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
requests-html>=0.10.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
python-dateutil>=2.8.0

# Frontend
gradio>=4.8.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
configargparse>=1.7

# Utilities
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
tenacity>=8.2.0
rich>=13.7.0

# Cloud API Integration
openai>=1.12.0
anthropic>=0.8.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Database (optional for caching)
sqlite3
sqlalchemy>=2.0.0

# Industry Classification
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Geographic Processing
geopy>=2.4.0
geonamescache>=2.0.0 